requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
google-genai>=1.0.0
google-cloud-firestore>=2.16.0
python-dotenv>=1.0.0
httpx>=0.27.0
yfinance>=0.2.36
numpy>=1.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
